+++ src/main/java/com/inventaire/inventaire/dao/LigneInventaireRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface LigneInventaireRepository extends JpaRepository<LigneInventaire, Long> {
+public interface LigneInventaireRepository extends JpaRepository<LigneInventaire, Integer> {
 
 }
